Provide environment variables to downstream builds
The cause would only be StashCause for the the top-level build of a
multi-configuration project.

First make sure it's an AbstractBuild. Non-build runs (e.g. git
invocations) should not need our environment variables (the variables are
substituted explicitly before the command is run).

Then find the root build and check it's cause.

Don't return prematurely to ensure that super.buildEnvironmentFor() is
run unconditionally.