Remove StashBuilds#getCause(), call build.getCause() directly
The "build" argument in onStarted() and onCompleted() is not of a raw
type anymore, which enables type-safe behavior without casts.

build.getCause() returns a value of the type provided in its argument. If
cause is null or there is no cause of the given type, build.getCause()
returns null.

That is the desired behavior. No extra wrapper is required.