StashBuildTrigger: Use isBuildable() instead of isDisabled()
There are reasons for a job to be non-buildable other than the project
being disabled. The Jenkins core uses isBuildable(), not isDisabled()
when deciding whether to run a job.

Instead of checking that stashPullRequestsBuilder is not null, check that
this.job is not null. Both are set in start(). Both would be null before
start() and non-null objects afterwards.

Use this.job instead of stashPullRequestsBuilder.getProject(). Both
values should be the same, as stashPullRequestsBuilder.getProject()
returns the value passed to the StashPullRequestsBuilder constructor.
That is the job argument to start(), which also becomes the value of
this.job.