StashRepository: Clean up comment sorting, add comments
When extracting parameters from comments, start with the oldest comments,
as the code iterates over all comments. Parameters from the new comments
are processed last, overwriting older values.

When deleting old "BUILD FINISHED" comments, don't sort the comments, it
doesn't matter in which order they are deleted.

When looking for comments to decide whether to run a build, start with
the most recent comments. The loop terminates early as soon as it's clear
whether to start the build or not. Don't sort and reverse the sort order
as two operations, use reverse sorting.