Merge StashBuilds into StashBuildListener
StashBuildListener was just a thin wrapper around StashBuilds, whereas
StashBuilds served as a utility class for StashBuildListener. It is not a
natural way to separate code into classes.

Remove StashPullRequestsBuilder#getBuilds(), which is not used anymore.
Add StashPullRequestsBuilder#getRepository() instead.

Add unit tests. The happy case for onCompleted() cannot be implemented
with Mockito at this time without additional code changes. Make sure that
onCompleted() is not triggered for a wrong cause.