Fill parameter values from environment variables added by the plugin
Parameters generate environment variables that take precedence over the
environment variables supplied by the EnvironmentContributor extension.

Some users relied on the plugin filling values for the parameters defined
for the project. That functionality was broken in version 1.9.

Make StashRepository#getParameters() fill parameters with the values that
the plugin supplies to the environment.

To avoid code duplication, move the environment variables to StashCause
as a transient map (the data is already saved to XML files from the
fields). Use that map both from StashBuildEnvironmentContributor and
StashRepository.

Make the environment variables override the parameters from the pull
request comments. Users with comment posting permissions should not be
able to override parameters that the administrator intended to be filled
from the environment.