StashPullRequestComment: Rewrite Comparable interface implementation
Define hashCode() and equals() to ensure consistent behavior and fix
FindBugs warnings. Remove FindBugs suppressions.

Treat null as Integer.MIN_VALUE in hashCode(), as that comment ID is not
expected to be found in actual server responses.

Make equals() and compareTo() compare ID only. Comparing text is not
needed because comparison is only used for comments in the same pull
request. Each comment in the pull request should have an ID unique within
that pull request.

Add extensive unit tests.