Add a "Polling Log" page as a link on the project page
Add StashPollingAction class, which is an Action and a logger. Use slf4j
MessageFormatter for the log messages. Add slf4j dependency.

StashPollingActionFactory is used to inject StashPollingAction into the
project using the Extension mechanism.

Change StashBuildTrigger and StashRepository to write information about
polling to the polling log. Some messages are now logged both to the
polling log and to the global Jenkins log.