StashPostBuildCommentAction: Remove, it has been obsolete since version 1.9
Starting with plugin version 1.9, this file has been used to suppress
warnings about unreadable data when upgrading from plugin version 1.8 and
older.

The compatibility code have been around for a long time, it should be
safe to remove now.

Not having this file in the version 1.13 of the plugin means that users
upgrading directly from version 1.8 and older directly to version 1.13
and newer would get warnings about unreadable fields. The warnings are
actually harmless and can be trivially cleared.

The unused fields were triggering warnings in SpotBugs and FindBugs when
threshold was set to Low.