StashBuildTrigger: Don't throw NullPointerException unnecessarily
If job is null, report it and return. Looking at the Jenkins code, it
should not happen, but Javadoc makes no explicit promises, and the job
argument is not annotated with @Nonnull.

Report all possible exceptions in start(), not just NullPointerException.
The calling code may not be well prepared to deal with exceptions.