StashApiClient: Don't ever return null from entityAsString()
EntityUtils.toString() can return null under some obscure circumstances.
From our point of view, it's the same as an empty response, as long as no
exception is thrown. JSON parser would consider both as invalid JSON, but
it's safer not to deal with null.

Use Util.fix to convert null to an empty string.

Mark getRequest() and postRequest() as @Nonnull, they return the result
of entityAsString().